pyproject.toml
[project]
name = "e20-co326-motor-health-monitoring-digital-twin-with-edge-anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
]
